+++ src/typechecker/stmnt.rs
@@ -1,9 +1,11 @@
-use crate::{ast, error::Error};
+use crate::{ast, error::Error, typechecker::var::Var};
 
 use super::{
     func::Func,
     scope::Scope,
+    types::Type,
     value::{FuncCall, GenericValue, Value},
+    var::{VarId, Vars},
     Typechecker,
 };
 
@@ -18,6 +20,7 @@ impl Return {
     pub fn check(
         checker: &Typechecker,
         scope: &mut Scope,
+        vars: &Vars,
         func: &Func,
         return_: &ast::Return,
     ) -> Result<Self, Error> {
@@ -34,9 +37,8 @@ impl Return {
                 });
             }
 
-            let value = GenericValue::check(scope, value)
-                .ok_or(Error::InvalidValue(value.span()))?
-                .coerce(&func.signature.returns.clone().unwrap())
+            let value = GenericValue::check(scope, vars, value)?
+                .coerce(vars, &func.signature.returns.clone().unwrap())
                 .ok_or(Error::InvalidReturnValue {
                     decl: func.span,
                     name: func.signature.returns.clone().unwrap().name(),
@@ -65,11 +67,59 @@ impl Return {
     }
 }
 
+/// A variable declaration.
+#[derive(Clone, Debug, Hash, PartialEq, Eq)]
+pub struct VarDecl {
+    pub var: VarId,
+    pub value: Option<Value>,
+}
+
+impl VarDecl {
+    pub fn check(
+        checker: &Typechecker,
+        scope: &mut Scope,
+        vars: &mut Vars,
+        func: &Func,
+        decl: &ast::Var,
+    ) -> Result<Self, Error> {
+        if decl.ty == None && decl.value == None {
+            return Err(Error::CannotInferVarType(decl.span));
+        }
+
+        let (ty, value) = if let Some(ty) = &decl.ty {
+            let ty = Type::check(scope, ty)?;
+
+            if let Some(value) = &decl.value {
+                let value = GenericValue::check(scope, vars, value)?
+                    .coerce(vars, &ty)
+                    .ok_or(Error::InvalidValue(value.span()))?; // TODO: make special variable value error
+
+                (ty, Some(value))
+            } else {
+                (ty, None)
+            }
+        } else {
+            let value =
+                GenericValue::check(scope, vars, &decl.value.clone().expect("Cannot be none"))?;
+
+            let value = value.coerce_default();
+
+            (value.ty(vars), Some(value))
+        };
+
+        let id = vars.declare_var(Var::new(decl.span, decl.name.value.clone(), ty));
+        scope.define_var(decl.name.value.clone(), id);
+
+        Ok(Self { var: id, value })
+    }
+}
+
 /// A statement of code.
 #[derive(Clone, Debug, Hash, PartialEq, Eq)]
 pub enum Stmnt {
     FuncCall(FuncCall),
     Return(Return),
+    VarDecl(VarDecl),
 }
 
 impl Stmnt {
@@ -77,19 +127,25 @@ impl Stmnt {
     pub fn check(
         checker: &Typechecker,
         scope: &mut Scope,
+        vars: &mut Vars,
         func: &Func,
         stmnt: &ast::Expr,
     ) -> Result<Self, Error> {
         match stmnt {
             ast::Expr::Call(call) => {
-                let func_call = FuncCall::check(checker, scope, call)?;
+                let func_call = FuncCall::check(checker, scope, vars, call)?;
 
                 Ok(Stmnt::FuncCall(func_call))
             }
             ast::Expr::Return(return_) => {
-                let return_ = Return::check(checker, scope, func, return_)?;
+                let return_ = Return::check(checker, scope, vars, func, return_)?;
 
                 Ok(Stmnt::Return(return_))
+            }
+            ast::Expr::Var(var) => {
+                let var = VarDecl::check(checker, scope, vars, func, var)?;
+
+                Ok(Stmnt::VarDecl(var))
             }
             _ => Err(Error::InvalidStatement(stmnt.span())),
         }
@@ -107,13 +163,14 @@ impl Block {
     pub fn check(
         checker: &Typechecker,
         scope: &mut Scope,
+        vars: &mut Vars,
         func: &Func,
         block: &ast::Block,
     ) -> Result<Self, Error> {
         let mut value = Vec::new();
 
         for stmnt in &block.value {
-            let stmnt = Stmnt::check(checker, scope, func, stmnt)?;
+            let stmnt = Stmnt::check(checker, scope, vars, func, stmnt)?;
             value.push(stmnt);
         }