+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - [feat (#5): [syntax/parser, typechecker, codegen] decimal integer literals](https://github.com/amp-lang/amp/issues/5)
 - [feat (#4): [syntax/parser, typechecker, codegen] return statements](https://github.com/amp-lang/amp/issues/4)
 - [feat (#8): [syntax/scanner, syntax/parser, typechecker, codegen] implement slice types](https://github.com/amp-lang/amp/issues/8)
+- [feat (#9): [syntax/scanner, syntax/parser, typechecker, codegen] implement variable declarations](https://github.com/amp-lang/amp/issues/9)
 
 ### Compiler
 - feat: Basic stopwatch benchmark for Amp compile times.